+++ pkg/vm_service/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.0
+- Support service protocol version 3.27:
+  - Added `getCpuSamples` and `clearCpuSamples` methods
+  - Added `CpuSamples`, `CpuSample`, and `ProfileFunction` classes.
+
 ## 1.1.2
 - Fixed issue where `closureFunction` and `closureContext` were only expected in
   `Instance` objects rather than `InstanceRef`.